+++ ci/data_extractor/src/formatters/common.py
@@ -1,5 +1,4 @@
 import collections
-import enum
 import pathlib
 import xml.dom.minidom
 from collections.abc import Callable
@@ -49,24 +48,6 @@ def compute_comparisons(*results):
 
 OPERATION_SIZE_COLUMN_HEADER = "Operation \\ Size"
 OPERATION_PRECISION_COLUMN_HEADER = "Operation \\ Precision (bits)"
-
-
-class OperationDisplayName(enum.StrEnum):
-    Negation = "Negation (-)"
-    AddSub = "Add / Sub (+,-)"
-    Mul = "Mul (x)"
-    EqualNotEqual = "Equal / Not Equal (eq, ne)"
-    Comparisons = "Comparisons (ge, gt, le, lt)"
-    MaxMin = "Max / Min (max, min)"
-    Bitwise = "Bitwise operations (&, \\|, ^)"
-    Div = "Div  (/)"
-    Rem = "Rem  (%)"
-    DivRem = "Div / Rem  (/, %)"
-    Shifts = "Left / Right Shifts (<<, >>)"
-    Rotates = "Left / Right Rotations (left_rotate, right_rotate)"
-    LeadingTrailing = "Leading / Trailing zeros/ones"
-    Log2 = "Log2"
-    Select = "Select"
 
 
 class BenchArray:
@@ -201,93 +182,12 @@ class GenericFormatter:
 
         :raises NotImplementedError: Raised when the specified layer is unsupported.
         """
-        match self.layer:
-            case Layer.Integer:
-                return self._format_integer_data(data, conversion_func)
-            case Layer.CoreCrypto:
-                return self._format_core_crypto_data(data, conversion_func)
-            case Layer.HLApi:
-                return self._format_hlapi_data(data, conversion_func)
-            case _:
-                raise NotImplementedError(f"layer '{self.layer}' not supported yet")
+        return self._format_data(data, conversion_func)
 
     @staticmethod
-    def _format_integer_data(data: dict[BenchDetails : list[int]], conversion_func):
-        formatted = collections.defaultdict(
-            lambda: {
-                2: "N/A",
-                8: "N/A",
-                16: "N/A",
-                32: "N/A",
-                64: "N/A",
-                128: "N/A",
-                256: "N/A",
-            }
-        )
-        for details, timings in data.items():
-            test_name = "_".join((details.sign_flavor.value, details.operation_name))
-            bit_width = details.bit_size
-            value = conversion_func(timings[-1])
-
-            if bit_width == 40:
-                # Ignore this width as it's not displayed publicly.
-                continue
-
-            formatted[test_name][bit_width] = value
-
-        return formatted
-
-    @staticmethod
-    def _format_core_crypto_data(data: dict[BenchDetails : list[int]], conversion_func):
-        params_set = set()
-        for details in data:
-            try:
-                params_set.add(details.get_params_definition())
-            except Exception:
-                # Might be a Boolean parameters set, ignoring
-                continue
-
-        params_set = sorted(params_set)
-
-        formatted = collections.defaultdict(
-            lambda: {params: "N/A" for params in params_set}
-        )
-        for details, timings in data.items():
-            try:
-                reduced_params = details.get_params_definition()
-            except Exception:
-                # Might be a Boolean parameters set, ignoring
-                continue
-
-            test_name = details.operation_name
-            value = conversion_func(timings[-1])
-            formatted[test_name][reduced_params] = value
-
-        return formatted
-
-    @staticmethod
-    def _format_hlapi_data(data: dict[BenchDetails : list[int]], conversion_func):
-        formatted = collections.defaultdict(
-            lambda: {
-                2: "N/A",
-                4: "N/A",
-                8: "N/A",
-                10: "N/A",
-                12: "N/A",
-                14: "N/A",
-                16: "N/A",
-                32: "N/A",
-                64: "N/A",
-                128: "N/A",
-            }
-        )
-        for details, timings in data.items():
-            test_name = details.operation_name.lstrip("ops::")
-            bit_width = details.bit_size
-            value = conversion_func(timings[-1])
-            formatted[test_name][bit_width] = value
-
-        return formatted
+    def _format_data(*args, **kwargs):
+        # Must be implemented by subclasses
+        raise NotImplementedError
 
     def generate_array(
         self,
@@ -322,348 +222,16 @@ class GenericFormatter:
 
         :raises NotImplementedError: If the current layer is not implemented.
         """
-        match self.layer:
-            case Layer.Integer:
-                return self._generate_unsigned_integer_array(
-                    data, operand_type, included_types, excluded_types
-                )
-            case Layer.CoreCrypto:
-                return self._generate_core_crypto_showcase_arrays(data)
-            case _:
-                raise NotImplementedError
-
-    def _generate_unsigned_integer_array(
-        self,
-        data,
-        operand_type: OperandType = None,
-        included_types: list[RustType] = ALL_RUST_INTEGER_TYPES,
-        excluded_types: list[RustType] = None,
-    ):
-        match operand_type:
-            case OperandType.CipherText:
-                prefix = "unsigned"
-            case OperandType.PlainText:
-                prefix = "unsigned_scalar"
-
-        match self.backend:
-            case Backend.CPU:
-                operations = [
-                    f"{prefix}_neg_parallelized",
-                    f"{prefix}_add_parallelized",
-                    f"{prefix}_mul_parallelized",
-                    f"{prefix}_eq_parallelized",
-                    f"{prefix}_gt_parallelized",
-                    f"{prefix}_max_parallelized",
-                    f"{prefix}_bitand_parallelized",
-                    f"{prefix}_div_rem_parallelized",
-                    f"{prefix}_left_shift_parallelized",
-                    f"{prefix}_rotate_left_parallelized",
-                    f"{prefix}_leading_zeros_parallelized",
-                    f"{prefix}_ilog2_parallelized",
-                    f"{prefix}_if_then_else_parallelized",
-                ]
-            case Backend.GPU:
-                operations = [
-                    f"{prefix}_neg",
-                    f"{prefix}_add",
-                    f"{prefix}_mul",
-                    f"{prefix}_eq",
-                    f"{prefix}_gt",
-                    f"{prefix}_max",
-                    f"{prefix}_bitand",
-                    f"{prefix}_div_rem",
-                    f"{prefix}_left_shift",
-                    f"{prefix}_rotate_left",
-                    f"{prefix}_leading_zeros",
-                    f"{prefix}_ilog2",
-                    f"{prefix}_if_then_else",
-                ]
-            case Backend.HPU:
-                operations = [
-                    f"{prefix}_sub",  # Negation operation doesn't exist in HPU yet
-                    (
-                        f"{prefix}_add"
-                        if operand_type == OperandType.CipherText
-                        else f"{prefix}_adds"
-                    ),
-                    (
-                        f"{prefix}_mul"
-                        if operand_type == OperandType.CipherText
-                        else f"{prefix}_muls"
-                    ),
-                    f"{prefix}_cmp_eq",
-                    f"{prefix}_cmp_gt",
-                    f"{prefix}_max",
-                    f"{prefix}_bw_and",
-                    (
-                        f"{prefix}_div"
-                        if operand_type == OperandType.CipherText
-                        else f"{prefix}_divs"
-                    ),
-                    (
-                        f"{prefix}_shift_l"
-                        if operand_type == OperandType.CipherText
-                        else f"{prefix}_shifts_l"
-                    ),
-                    (
-                        f"{prefix}_rot_l"
-                        if operand_type == OperandType.CipherText
-                        else f"{prefix}_rots_l"
-                    ),
-                    f"{prefix}_lead0",
-                    f"{prefix}_ilog2",
-                    f"{prefix}_if_then_else",
-                ]
-            case _:
-                raise NotImplementedError(
-                    f"backend '{self.backend}' not supported yet for integer formatting"
-                )
-
-        display_names = [
-            OperationDisplayName.Negation,
-            OperationDisplayName.AddSub,
-            OperationDisplayName.Mul,
-            OperationDisplayName.EqualNotEqual,
-            OperationDisplayName.Comparisons,
-            OperationDisplayName.MaxMin,
-            OperationDisplayName.Bitwise,
-            OperationDisplayName.DivRem,
-            OperationDisplayName.Shifts,
-            OperationDisplayName.Rotates,
-            OperationDisplayName.LeadingTrailing,
-            OperationDisplayName.Log2,
-            OperationDisplayName.Select,
-        ]
-
-        types = included_types.copy()
-        excluded_types = excluded_types if excluded_types is not None else []
-        for excluded in excluded_types:
-            types.remove(excluded)
-
-        first_column_header = OPERATION_SIZE_COLUMN_HEADER
-
-        # Adapt list to plaintext benchmarks results.
-        if operand_type == OperandType.PlainText and self.backend != Backend.HPU:
-            if self.backend == Backend.CPU:
-                div_name = f"{prefix}_div_parallelized"
-                rem_name = f"{prefix}_rem_parallelized"
-            elif self.backend == Backend.GPU:
-                div_name = f"{prefix}_div"
-                rem_name = f"{prefix}_rem"
-
-            operations.insert(8, div_name)
-            operations.insert(9, rem_name)
-            operations.pop(7)
-
-            display_names.insert(
-                8,
-                OperationDisplayName.Div,
-            )
-            display_names.insert(
-                9,
-                OperationDisplayName.Rem,
-            )
-            display_names.pop(7)  # Remove Div / Rem
-
-            # Negation operation doesn't exist in plaintext
-            operations.pop(0)
-            display_names.pop(0)
-
-        data_without_excluded_types = {}
-        for op, values in data.items():
-            try:
-                data_without_excluded_types[op] = {
-                    typ: val
-                    for typ, val in values.items()
-                    if RustType.from_int(typ) in types
-                }
-            except NotImplementedError:
-                # Unknown type from database, ignoring
-                continue
-
-        filtered_data = filter(lambda t: t in operations, data_without_excluded_types)
-        # Get operation names as key of the dict to ease fetching
-        filtered_data_dict = {
-            item: tuple(data_without_excluded_types[item].values())
-            for item in filtered_data
-        }
-
-        result_lines = []
-        for name, op in zip(display_names, operations):
-            try:
-                line = {first_column_header: name.value}
-                line.update(
-                    {
-                        types[i].name: value
-                        for i, value in enumerate(filtered_data_dict[op])
-                    }
-                )
-                result_lines.append(line)
-            except KeyError:
-                # Operation not found in the results, ignoring this line.
-                print(
-                    f"backend '{self.backend}' could not find operation '{op}' to put in line '{name}'"
-                )
-                continue
-
-        return [
-            BenchArray(result_lines, self.layer),
-        ]
-
-    def _build_results_dict(
-        self,
-        pfails: list[ErrorFailureProbability],
-        noise_distributions: list[NoiseDistribution],
-        operation_displays: list[CoreCryptoOperation],
-        default_precisions: Callable[[], dict],
-    ):
-        results_dict = {}
-
-        for pfail in pfails:
-            for noise in noise_distributions:
-                results_dict[CoreCryptoResultsKey(pfail, noise)] = {
-                    o: default_precisions() for o in operation_displays
-                }
-
-        return results_dict
-
-    def _generate_core_crypto_showcase_arrays(
-        self,
-        data,
-    ):
-        supported_pfails = [
-            ErrorFailureProbability.TWO_MINUS_64,
-            ErrorFailureProbability.TWO_MINUS_128,
-        ]
-        noise_distributions = [
-            NoiseDistribution.Gaussian,
-            NoiseDistribution.TUniform,
-        ]
-
-        sorted_results = self._build_results_dict(
-            supported_pfails,
-            noise_distributions,
-            OPERATIONS_DISPLAYS,
-            DEFAULT_CORE_CRYPTO_PRECISIONS,
+        return self._generate_arrays(
+            data,
+            operand_type,
+            included_types=included_types,
+            excluded_types=excluded_types,
         )
 
-        for operation, timings in data.items():
-            try:
-                formatted_name = CoreCryptoOperation.from_str(operation)
-            except NotImplementedError:
-                # Operation is not supported.
-                continue
-
-            for param_definition, value in timings.items():
-                pfail = param_definition.p_fail
-                if pfail not in supported_pfails:
-                    print(f"[{operation}] P-fail '{pfail}' is not supported")
-                    continue
-                noise = param_definition.noise_distribution
-                precision = int(param_definition.message_size) * 2
-                key = CoreCryptoResultsKey(pfail, noise)
-
-                if (
-                    formatted_name == CoreCryptoOperation.MultiBitPBS
-                    or formatted_name == CoreCryptoOperation.KeySwitchMultiBitPBS
-                ) and param_definition.pbs_kind != PBSKind.MultiBit:
-                    # Skip this operation since a multi-bit operation cannot be done with any other parameters type.
-                    continue
-                elif (
-                    formatted_name == CoreCryptoOperation.PBS
-                    or formatted_name == CoreCryptoOperation.KeySwitchPBS
-                ) and param_definition.pbs_kind != PBSKind.Classical:
-                    # Skip this operation since a classical operation cannot be done with any other parameters type.
-                    continue
-
-                grouping_factor = param_definition.grouping_factor
-                if (
-                    grouping_factor is not None
-                    and grouping_factor != self.requested_grouping_factor
-                ):
-                    continue
-
-                if param_definition.details["variation"] not in ["", "BENCH"]:
-                    continue
-
-                try:
-                    sorted_results[key][formatted_name][precision] = value
-                except KeyError:
-                    # Operation is not supposed to appear in the formatted array.
-                    continue
-
-        first_column_header = OPERATION_PRECISION_COLUMN_HEADER
-
-        arrays = []
-        for key, results in sorted_results.items():
-            array = []
-            for operation, timings in results.items():
-                d = {first_column_header: operation.value}
-                d.update({str(k): v for k, v in timings.items()})
-                array.append(d)
-
-            arrays.append(
-                BenchArray(
-                    array,
-                    self.layer,
-                    metadata={
-                        "noise": key.noise_distribution,
-                        "pfail": key.pfail,
-                    },
-                )
-            )
-
-        return arrays
-
-
-# ---------------------------
-# Core_crypto layer constants
-# ---------------------------
-
-OPERATIONS_DISPLAYS = [
-    # CoreCryptoOperation.KeySwitch, # Uncomment this line to get keyswitch in the tables
-    CoreCryptoOperation.PBS,
-    CoreCryptoOperation.MultiBitPBS,
-    CoreCryptoOperation.KeySwitchPBS,
-    CoreCryptoOperation.KeySwitchMultiBitPBS,
-]
-
-DEFAULT_CORE_CRYPTO_PRECISIONS = lambda: {
-    2: "N/A",
-    4: "N/A",
-    6: "N/A",
-    8: "N/A",
-}
-
-
-class CoreCryptoResultsKey:
-    """
-    Representation of a hashable result key for the core_crypto layer.
-
-    :param pfail: Probability of failure associated with the cryptographic result.
-    :type pfail: ErrorFailureProbability
-    :param noise_distribution: Noise distribution parameter linked to the
-        cryptographic result.
-    :type noise_distribution: NoiseDistribution
-    """
-
-    def __init__(
-        self, pfail: ErrorFailureProbability, noise_distribution: NoiseDistribution
-    ):
-        self.pfail = pfail
-        self.noise_distribution = noise_distribution
-
-    def __eq__(self, other):
-        return (
-            self.pfail == other.pfail
-            and self.noise_distribution == other.noise_distribution
-        )
-
-    def __hash__(self):
-        return hash((self.pfail, self.noise_distribution))
-
-    def __repr__(self):
-        return f"CoreCryptoResultsKey(pfail={self.pfail}, noise_distribution={self.noise_distribution})"
+    def _generate_arrays(self, *args, **kwargs):
+        # Must be implemented by subclasses
+        raise NotImplementedError
 
 
 class CSVFormatter(GenericFormatter):
