+++ ci/data_extractor/src/formatters/integer/__init__.py
@@ -1,0 +1,1 @@
+from .integer import *
